feat: Support reading from `stdin` stream in `invoke` and `endpoint` commands
Some restructuring allows us to stay largely oblivious to the kind of invocation
the function has been subjected to.

We also want to embrace the convention that middleware in `lib.js` can be
plugged into a pipeline at any stage by mutating the `ctx` argument in place
and handing that one down instead of assuming that the current stage is the
very first or last one. We'll see how this works out, user code should typically
avoid mutating arguments for the sake of personal sanity.