feat: Convert `invoke` and `serve` to scripts inside the `fn/` directory
Right now it feels appealing to provide the same API for all runtimes in the
form of `fn/{invoke,serve}`. The specifics would then be handled by the actual
runtime library without the instrumentation scripts needing to know which one
is being used. We'll see how this turns out when we have another runtime to
play with.