chore: Remove obsolete `fn/runtime/nodejs-lts` directory
It's not yet clear how to best distribute the different runtimes. The current
thinking is that we just provide all the core runtimes within the `fn` directory
as source code and have the `fn/<cmd>` commands figure out which one to use
based on some configuration. That might be overkill in terms of wasting space
but it would enable very easy updates via `git pull` circumventing all kinds
of versioning problems while retaining straight forward customization by the
users themselves. Maybe the user can just use the `.dockerignore` to not bloat
the docker image?

Sidenote
========

As the user would basically distribute the source code of all the runtimes
we will have to think hard about licensing at some point.