feat: Introduce `invoke` Context API to call functions
This is just a toy implementation that only knows about endpoints within the
same container. It illustrates how the API could look and what steps are
involved in getting the functions independent of each other via message passing.

We also see that the runtime is accumulating ever more cruft and given that this
is only scratching the surface without the hard edge cases this is clearly not
something we want to re-implement for every runtime in the future.